requirements.txt
PyYAML==6.0.1
colorlog==6.7.0
Flask==2.0.3
python-gitlab==2.10.1